+++ src/gsUtils/gsMesh/gsMesh.h
@@ -54,9 +54,9 @@ public:
 
         // copy all pointers to it's original counterpart in face
         cloneAll(mesh.face, face);
-        for (int i = 0; i < mesh.face.size(); ++i)
+        for (size_t i = 0; i < mesh.face.size(); ++i)
         {
-            for (int j = 0; j < 3; ++j)
+            for (size_t j = 0; j < 3; ++j)
             {
                 GISMO_ASSERT(vertex[mesh.face[i]->vertices[j]->getId()]->getId() == mesh.face[i]->vertices[j]->getId(), "gsMesh(const gsMesh<T> & mesh): getId() of vertex and face don't match");
                 face[i]->vertices[j] = vertex[mesh.face[i]->vertices[j]->getId()];
@@ -65,7 +65,7 @@ public:
 
         // iterate over all edges and make them new
         edge = mesh.edge;
-        for (int i = 0; i < mesh.edge.size(); ++i)
+        for (size_t i = 0; i < mesh.edge.size(); ++i)
         {
             GISMO_ASSERT(mesh.edge[i].source->getId() == vertex[i]->getId(), "gsMesh(const gsMesh<T> & mesh): getId() of vertex and edge.source don't match");
             edge[i].source = vertex[mesh.edge[i].source->getId()];
